requirements.txt
Requests
cachecontrol
PyYAML
dotmap
PyGitHub
python-gitlab
travispy
nltk
graphqlclient
docutils
jsonschema
beautifulsoup4
tqdm